requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
earthengine-api==0.1.384
pandas==2.1.4
requests==2.31.0
geopandas==0.14.1
python-dotenv==1.0.0
geopy==2.4.1